Extract methods
